refactor(portable_dropbot_ui): consolidate panes into More/Advanced
Fold Temp & Lighting and PMT into one More Portable Dropbot
Controls pane (collapsible Temperature and PMT groups; heater
start/stop is one toggle, confirm/read actions are compact glyph
buttons, the acquire hint word-wraps) and Power System + Motor
Params into one advanced-mode-locked Advanced Portable Dropbot
Controls pane. The raw lighting controls are removed entirely (the
status pane owns lighting). Four listeners become two; merged MVC
modules keep history via git renames.
